fixes RangeEditor not properly instantiated (#29, #97)
Because the class attribute was overwritten instead of only
the green value being replaced, sometimes an Editor would
be instantiated instead of RangeEditor on subsequent edits.
The behavior would depend on when the AngularJS 2-way binding
would kick in and "fix" the class attribute.

Since Gold is set to 6, outside the linear scale for green
values (0 to 3), the average value could end-up to 5 when
one of the value summary fields would have been recorded
(incorrectly) to 4 (ex: (6+4)/2).