aggregates MyTSP dashboard based on rollup_scores
Before this commit, MyTSP dashboard uses the djaodjin-survey
Matrix/Metric implementation to aggregate scores multiple ways.
This has proven too slow for on-the-fly computations, even with
a small number of organizations.
Since at this point all the cohorts are following the industry
segment sections, we rely on the same algorithms used to display
scorecard (with late adjustments).