fixes nb_respondents showing on improvement plan
Because there was no "Yes" answers, the inner join
between yes_view and yes_no_view would be empty
and nb_respondents defaulted to zero.
We do a left outer join with yes_no_view on the left
side to keep the number of respondents regardless
of the actual answers.