fixes ".labels .label" to use webfont (#138)
After cleaning up paths to load URLs from the local filesystem,
and only use the css required to remove warning messages from Weasyprint,
it turned out that the first item in font-family was mispelled,
getting the web browser to use Helvetica. Of course Helvetica is not
available on Linux (i.e. where the PDF is rendered).