shows historical samples in dashboard
Special note:
- replaced `WHERE survey_answer.id IS NULL OR survey_answer.metric_id = 2`
in favor of `WHERE survey_answer.metric_id = 2` in api/dashboards.py to
account for questions that were added after a score was frozen.
This has the effect of giving a score to organization that click complete
without answering all questions. We prevent that situation to happen
by other means (ex: disable "complete" button).