custom permissions check to access uploaded supporting documents
The assets/ upload API returns a fully qualified URL that is then
stored as a textual note in the survey sample (i.e. `Answer` with
unit='freetext' referencing a `Choice`).
The fully qualified URL contains the slug of the profile the document
belongs to. We thus cannot play the same trick as done to access
the Sample itself (see `mixins.ReportMixin`).